requirements.txt
GDAL==3.1.4
h5py==2.9.0
matplotlib==3.3.4
netCDF4==1.5.8
numba==0.45.1
numpy==1.16.2
ogr==0.32.0
opencv_contrib_python==4.5.4.58
osgeo==0.0.1
pandas==0.25.1
pyshp==2.1.3
python_dateutil==2.8.2
scikit_image==0.15.0
Shapely==1.7.1
skimage==0.0
